Worms Clone
I am aware of an issue within the latest commitment of this project, I
intend to resolve this tomorrow. This project is my attempt at creating
a 'clone' of a combination of the Worms games (by Team17), for testing a
few ideas I have in mind, as well as a 'proof of confidence' to myself.