Dig For Diamonds
This game was also produced at the University of Huddersfield (during
the 2011/2012 academic year). For this project though, I used the
Eclipse IDE and Java. Once again, I am sorry if one is not able to get
this project working with that IDE or other solutions, if Eclipse is not
available for use.